refactor(Sprint5): main 코드 정리
